requirements.txt
# Core FastAPI dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2  # Explicit bcrypt version to avoid version detection issues
python-multipart==0.0.12

# Payment Gateway (Paddle - PRIMARY)
# Note: Paddle SDK - use paddle-python-sdk or integrate via REST API
# paddle-python-sdk==0.2.1  # Uncomment when ready to use
# For now, we'll use httpx to call Paddle REST API directly

# HTTP Client (for calling zola-lead API and Paddle API)
httpx==0.27.2

# Caching
redis==5.2.0
hiredis==3.0.0

# Environment & Configuration
python-dotenv==1.0.1

# Email (for notifications)
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Monitoring
sentry-sdk[fastapi]==2.0.0

# Code Quality (optional)
black==24.10.0
flake8==7.1.1
mypy==1.13.0
